requirements.txt
open_clip_torch
huggingface_hub
tokenizers
opencv-python-headless